% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCandidates.R
\name{findCandidates}
\alias{findCandidates}
\title{Identify Regions of Interest (ROIs) for trace ions bleeding through large RT regions}
\usage{
findCandidates(files, dmz = 0.002, ...)
}
\arguments{
\item{files}{character vector with pathnames to mzML files}

\item{dmz}{mz allowance to merge ROIs from multiple files}

\item{...}{Pass on arguments to getROI()}
}
\value{
Data frame with candidates
}
\description{
Finds and merges ROIs from multiple files
}
\examples{
candidates <- findCandidates(mzMLfiles)
}
