% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getROI.R
\name{getROI}
\alias{getROI}
\title{Identify Regions of Interest (ROIs) for trace ions bleeding through large RT regions}
\usage{
getROI(file, prefilter = c(400, 2000), noise = 500, ppm = 40)
}
\arguments{
\item{file}{mzML to scout for ROIs}

\item{prefilter}{Criterion for ROI detection: number of scans with intensity over a limit (defaults to c(400, 2000))}

\item{noise}{Noise level (defaults to 500)}

\item{ppm}{Between-scan ppm allowance (defaults to 40)}
}
\value{
Data frame mz, scans, etc
}
\description{
Using findmzROI from XCMC, author: Johannes Rainer
}
\examples{
getROI(mzMLfile)
}
