% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzRecal.R
\name{mzRecal}
\alias{mzRecal}
\title{Recalibrate mass axis for MS1 (mzML format) - One file only}
\usage{
mzRecal(
  file,
  mzCandidates,
  ppmFind = 50,
  ppmQuantile = 0.2,
  ppmLimit = 5,
  intensityLimit = 1000,
  calibLimit = 300,
  preCalibrate = TRUE,
  twoStage = TRUE,
  ppmFindRough = 100,
  ppmQuantileRough = 0.1,
  plot = TRUE,
  jpg = TRUE,
  save = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{file}{mzML file}

\item{mzCandidates}{data frame}

\item{ppmFind}{ppm limit to discover candidate from data}

\item{ppmQuantile}{ppm quantile limit to exclude potential candidates from calibration (deviation from observed ppms)}

\item{ppmLimit}{ppm quantile limit to exclude potential candidates from calibration (mean absolute error)}

\item{intensityLimit}{minimum signal intensity (to avoid "frayed rope")}

\item{calibLimit}{Lower limit, above which there must be a candidate in order to perform calibration}

\item{preCalibrate}{If TRUE, performs a pre-calibration of the mass axis using the calibration line from the last scan}

\item{twoStage}{If TRUE, performs a rough calibration in-between pre-calibration (using last scan settings) and fine calibration}

\item{ppmFindRough}{ppm limit to discover candidate from data in rough calibration (if twoStage = TRUE)}

\item{ppmQuantileRough}{ppm quantile limit to exclude potential candidates from rough calibration}

\item{plot}{If TRUE plots calibration metrics (mz and number of calibrants; intercept and slope)}

\item{jpg}{If TRUE and plot = TRUE plots to jpg file in "mzRecal_log/" folder}

\item{save}{If TRUE saves recalibrated file in "mzRecal/" folder}

\item{verbose}{If TRUE reveals additional verbose information (if save = TRUE, sinks to txt file in "mzRecal_log/" folder)}

\item{...}{}
}
\value{
saves recalibrated file
}
\description{
Please see mzRecalibrate for standard interface for multiple files and parallelization
}
\examples{
# To be added
}
