% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzRecalibrate.R
\name{mzRecalibrate}
\alias{mzRecalibrate}
\title{Recalibrate mzML files}
\usage{
mzRecalibrate(
  files,
  mzCandidates,
  ppmFind = 50,
  ppmQuantile = 0.2,
  ppmLimit = 5,
  intensityLimit = 1000,
  calibLimit = 300,
  twoStage = TRUE,
  preCalibrate = TRUE,
  ppmFindRough = 100,
  ppmQuantileRough = 0.1,
  plot = TRUE,
  jpg = TRUE,
  save = TRUE,
  verbose = TRUE,
  parallel = TRUE,
  ...
)
}
\arguments{
\item{files}{mzML file names (character vector)}

\item{mzCandidates}{data frame}

\item{ppmFind}{ppm limit to discover candidate from data}

\item{ppmQuantile}{ppm quantile limit to exclude potential candidates from calibration (deviation from observed ppms)}

\item{ppmLimit}{ppm quantile limit to exclude potential candidates from calibration (mean absolute error)}

\item{intensityLimit}{minimum signal intensity (to avoid "frayed rope")}

\item{calibLimit}{Lower limit, above which there must be a candidate in order to perform calibration}

\item{twoStage}{If TRUE, performs a rough calibration in-between pre-calibration (using last scan settings) and fine calibration}

\item{preCalibrate}{If TRUE, performs a pre-calibration of the mass axis using the calibration line from the last scan}

\item{ppmFindRough}{ppm limit to discover candidate from data in rough calibration (if twoStage = TRUE)}

\item{ppmQuantileRough}{ppm quantile limit to exclude potential candidates from rough calibration}

\item{plot}{If TRUE plots calibration metrics (mz and number of calibrants; intercept and slope)}

\item{jpg}{If TRUE and plot = TRUE plots to jpg file in "mzRecal_log/" folder}

\item{save}{If TRUE saves recalibrated file in "mzRecal/" folder}

\item{verbose}{If TRUE reveals additional verbose information (if save = TRUE, sinks to txt file in "mzRecal_log/" folder)}

\item{parallel}{If TRUE and backend is registered (doParallel), will perform calculations in parallel}

\item{...}{Additional arguments}
}
\description{
Distributes recalibration to multiple threads if a parallel backend is registered
}
\examples{
# To be added
}
